requirements.txt
django
djangorestframework
djangorestframework-simplejwt
# psycopg2
python-dotenv
psycopg2-binary


